refactor things conditional on loopless into separate files
